Correct N2O data for O17

This function takes d45 and d46 and corrects the values for O17 to derive raw d15 and d18 values. The equations used are based on those derived by Jan Kaiser and Thomas Röckmann in "Correction of mass spectrometric isotope ratio measurements for isobaric isotopologues of O2, CO, CO2, N2O and SO2" (Rapid Communications in Mass Spectrometry, 2008, 3997--4008). It is important to note that this function does not currently take site preference into consideration

correct_N2O_for_17O(
  data,
  d45,
  d46,
  ref_17R = 0.0003799,
  ref_18R = 0.0020052,
  ref_15R_avg = 0.0036782,
  lambda = 0.52,
  d_max = 1000,
  quiet = FALSE
)

Arguments

data
the data frame

d45
column (in permil)

d46
column (in permil)

ref_17R
the 17O/16O reference ratio, VSMOW by default (value from Baertschi, 1976)

ref_18R
the 18O/16O reference ratio, VSMOW by default (value from Li et al., 1988)

ref_15R_avg
the 15N/14N reference ratio for the average of both alpha and beta N, N2 air by default (value from De Bievre et al., 1996)

lambda
the mass dependent scaling coefficient for the oxygen isotopes, default 0.52 (value from Werner and Brandt, 2001)

d_max
the maximum +/- delta value to consider in the root finding [in permil], should not need to change this unless samples are heavily enriched

quiet
whether the function should output information messages or be quiet (default is to output)

Value

the data frame with corrected 17O expects raw delta values (i.e. NOT in permil) expects raw delta values (i.e. NOT in permil) expects raw delta values (i.e. NOT in permil)